Updated ark-demo in `README.md`
The `README.md` filed contained a small demo scenario. The scenario was
somewhat out-dated and needed an update.

I've reworked the `run_bitcoind.sh` script and renamed it to
`ark_demo.sh`. Instead of starting `bitcoind` it creates a demo
environment useful aliases.

All bash lines started with `$`. It looks prettier but I decided to
remove it in favour of mindless copy paste.