Add integration tests
This commit contains a test for `bark-client` and `bark-aspd`.

I have to admit the current tests are somewhat pointless. But they are
useful for demonstrating the underlying purpose.

It should be easy to run the tests and swap the executable under test.
When updating the server, we should ensure that our clients don't have
to update immediately.

Users can set `ARKD_EXEC` or `NOAH_EXEC` to set the executable under
test. By default we use `cargo run --package arkd --` to ensure the
executable always matches the version in our workdir.