Create `TestContext`
The `TestContext` can be initialized in each test. The context creates a
directory which might be cleaned up once the test is completed. If you
want to leave folder intact set `TEST_LEAVE_INTACT=1`.

Currently, we just create a folder. In the next commits we will write
data such as database states and logs to that directory. This allows
developers to inspect the full state when investigating failing tests.