Change directory structure for test-data
Previously all test-data was stored in the `/tmp`-directory.
I have made the behavior configurable and changed the default to the
`./test`-folder in the workspace root.

I've also modified how `TestContext` works. The test-data is left intact
by default. It is also easier to find the test-data because we don't use
randmly generated folder names by default.

The only downside is that weird things might happen if you do multiple
evocations of `cargo test` concurrently. I'll assume nobody will be
doing that.