Previously, I usec `cargo run --package bark-aspd --` to run the `aspd`.
This works great if all depedencies are already compiled. HOwever, when mutliple tests are run in parallel this results in awkward locks on the build-directory.

It was a bad idea to use this by default and this option should be removed.

We now have a just-file.
This just-file will ensure all binaries are compiled before running the tests. The just-file also sets an environment variable that ensures the tests can find the `aspd` and `bark` executable
Don't compile in tests