ark-testing: Test bark create fails
Create a test in which the `bark create` command fails.
The main goal is to test if the command is atomic

Either the command should succeed and the folder should be created
The command should fail and no folder should be created.

If the folder elready exists it should raise an error.
Except if `--force` is used the folder should be overwritten