bark-cln: Subscribe to updates in sendpay
The `pay`-command uses `sendpay` under the hood.
When `pay` is triggered the command will parse the invoice and try to
find a route. Using that route `sendpay` is triggered which tries to
execute the payment.

The `pay` can be pretty clever. It will retry payments over a different
route or split the payment into parts and call `sendpay` multiple times.

It is commonly believed the lightning payments are fast. But this is not
always the case. Some routing-nodes can be slow and your payment might
be pending for a while. If the route provided to `sendpay` doesn't work
the `pay`-command will try to execute the payment over a different
route.

One downside of the pay-command is that it doesn't give any feedback
when the payment is ongoing. This can be frustrating for users who think
their program might be stuck. When the `pay`-command is used over `grpc`
it will often result in time-outs. The user will get an error and might
falsely beleive their payment failed.

To alleviate this problems we want to see when `sendpay` is called.
It appears their is no subscription for all `sendpay` events. That is
why I've implemented some custom method using the `wait` infrastructure.

The idea is pretty simple.

We use the `wait` command to ensure we our notified whenever an update
happens. If an update is seen we call `listsendpays` with some index
arguments. This allows us to see the updated values.