aspd: Use streams for payments
We noticed some payments take long which leads to the following problems
1. Time-outs
2. Users get no feedback

We modify the `FinishBbolt11Payment` to return a stream.
The first advantage is that streams don't time out.
The second advantage is that we can give status-updates.
Users are able to see that we're on it.

I've created `FinishBolt11Payment2`.
Don't worry, i'll remove the old one and rename this one in a later commit

However, having both versions at the same time will help us to
switch-over step-by-step. We can do multiple small commits and keep all
of them working.