Doublecheck pay command
The pay-command is tricky.
If it fails we double-check the payment-status.

If it is pending we keep checking, otherwise we return.

I've also moved the logic to the `lightning.rs`-file.
I do beleive it is cleaner. But also not that we are not taking
`&self` as an argument anymore.

It is now possible to launch the `pay_bolt11`-command in a separate task
without any `'static`-lifetime concerns