Ark-testing: Consistently use async everywhere
I've decided to use tokio::process::Command consistently.

We have 2 requirements
1. Logs are reliably captured in their file
2. Handlers are notified

Previously, we routed the command stdout into a pipe().
In this scenario it is easy to implement the handlers. However,
you have to manually ensure all logs are written to the file.

This forced us into a sync approach because the `Drop` trait doesn't
support async. When a `Daemon` is dropped we waited until the thread
that processed the log-files was completed.

We know switch the approach around. We route stdout directly to the file
which is very reliable. We are now processing the logs by reading the
file. The waiting a 100ms is a bit weird. But the approach makes things
more consistent.