Ark-testing: Make the bitcoin rpc time-out configurable
By default, the jsonrpc crate creates a socket with a 15 second timeout.

This is insuficcient when `generatetoaddress` is used with a high number
of blocks. In that case the socket times out and a SocketError(OS( code
35)) is returned.

This commit introduces the BITCOINRPC_TIMEOUT_SECS environment variable.
The developer can configure the value to their needs