Make TxStatus private and rename it
I renamed the enum from `TxExitStatus` to `TxStatus`.
It is already in the `exit`-module so the context should be sufficiently
clear.

I also decided to make this struct private. We don't need an API to
query the status of each individual transactions. Users already can
access that data through a mempool-explorer, I don't want to commit to
keeping this API stable.