ark-testing: Use lightning docker image on linux
Run the docker-container as the current user. I've also moved all data
from the `/root` to the `/data` directory to ensure the current user can
always access it.

This commit is required to make test work on my debian machine.
Otherwise, all mounted directories would be owned by root. The testing
infrastructure wouldn't be able to read the directories which prevents
us from reading the SSL-certificates