bark: Recover from unregistered board
If the client executes a boards but fails to register it to the server
we used to enter a broken state. The client has a vtxo but the server
wouldn't allow usage.

We are now ensuring the client will retry to register the board to the
server whenver the `maintenance`-command is called.