Give developer option to pick root-certificates
Nittesh B found an issue that it is not possible to use bark on android
(yet). We have to play with the tls certificates to make it work.

This is an alternative to his fix. The core-idea is that I want to give
the app-developer maximal control on which certificates are used.