bark: handle vtxos that are currently used in a non-finalised round
In this PR, we're introducing the notion of in-round VTXOs: round finalisation is not synchronous anymore so there is a gap between when VTXOs are provided as inputs and when they are actually marked as spent. To ensure they are not double-spent by mistake during this span, they are marked as `in-round`. A dedicated balance has been added in the `offchain_balance` API.