aspd: Force status in all times in the TxIndex
Previously, when a Transaction was added to the TxIndex it would not
have a status. On the next run of TxIndexProcess a status would be
assigned.

A process that called `status_of` would poll the status until
TxIndexProcess assigned a status. This has proven to be buggy as seen in
https://codeberg.org/ark-bitcoin/bark/issues/641

I am now forcing each transaction to have a status.
This means that the user has to either provide a status when inserting a
transaction to the cache or must provide a bitcoind that can request the
status.

The downside of providing a bitcoind is that the method becomes
fallible. It will return a Result.

It appears that both the Sweeper and ForfeitWatcher handle the error
cases correctly.

ForfeitWatcher remembers all `new_forfeits` and will only remove them
once we succeed to adding them to the index.

The Sweeper will just try to add the transaction to the index on the
next Sweep attempt.