ark-testing: use time-out more sparesely on bark
The BarkD-struct has a time-out. By default, we stop trying if a command
didn't succeed within the allotted time-frame.

In some tests we rely on a time-out to verify that we cannot perform a
refresh. To make the tests run faster we have reduced the time-out.

However, when testing under high load tests became flakey. Other
commands failed randomly because they cannot hit the 3.5 second
deadline.

I've moved the timeout to ensure we only apply the short time-out on the
commands that actually need it. I've also increased the time-out. This
ensures we actually see the expected errors before timing out.