testing: Move, rename and refactor cleanup_dbs
Moved it to the new postgres module.

The previous function allowed us to drop all databases with a similar
name. This is overkill and we aren't using that functionality so I
did a refactor to make it shorter.

I've also renamed the function to `drop_database` to clarify the
intention.