testing: Create a script that dumps the aspd postgres schema to stdout
I opted to put the script in ark-testing.

At first glance it seemed sensible to put the script in `aspd`.
This would be in line with the script that dumps the sqlite-schema for
bark.

This turned out to be a little bit tricky. To run the script we must
configure and launch a postgres database. We already have all this code
available in ark-testing but we cannot touch it in the `aspd`-crate.

I opted to put the script in ark-testing instead