aspd: Refactor TxIndex with cache
We add TxIndexData which holds the index's data.
The TxIndexData is just a plain structure to hold data.
It has nearly no fields and is designed to be testable.

The TxIndexProcess will do the updates.

The TxIndex will ensure that each transaction
is pushed to the database. It will also retrieve
things from the database if required.

We use weak references in TxIndexData

This should help us to use less memory.
Each bitcoin::Transaction that isn't needed anymore will
be erased from memory.