bark: Replace closure-based round participation with struct
This was a tough decision. The closure is really flexible.
But it's just not workable to store a closure in a state,
so we can't store a round state before the round started.

I have some ideas of how we can allow users to provide a custom
`RoundFunder` through a trait and provide their own
`DesiredRoundParticipation::Custom` value etc for lateron doing
more fancy stuff. For now, this works.