bark: Rename spendable to claimable in the context of an exit
We have the concept of an `ExitVtxo` which is a `Vtxo` going through an
exit.

When going exiting a VTXO the client has to
1. bring all vtxos onchain
2. wait until some delta expires
3. claim the exit by spending it

When the delta was expired we called the ExitVtxo spendable.
I tought this was confusing because a Vtxo can be spendable as well.

I am now using Claimable in the context of an exit