protocol: Remove funding_tx from RegisterBoard
This is a follow-up from #1028

Previously, the protocol required a board_tx in the
`register_board`-call. This is required to validate
the vtxo.

However, we have changed the behavior. Now the board_tx
needs to have sufficient confirmations before the server
will accept the registration.

The server doesn't need to receive the board_tx form the user anymore
and can receive it from chain.

Apparantly, we never edited the protocol in #1028.

PS: This has a nice side-effect. It also cleans up the
bark API significantly because we will not need any onchain
reference in sync.