testing: Reproduce bug in sweeper
This is a repro of https://codeberg.org/ark-bitcoin/bark/issues/1165.

It might be useful to add a little bit of context.
The root cause of the bug is when we query the database for boards.

We do this by querying each VTXO for which the funding_txid is not a
round.

If you create a board and send it out_of_round you will create 2 new
vtxos. The server database will create the following vtxos
1. The old VTXO (spent)
2. The new VTXO (unspent)
3. The change VTXO (unspent)

Note, that all of these 3 VTXOs originate from the same funding_tx.

This means the sweeper will try to seep the same outpoint 3 times.
Which is the root cuase of the panic