server: Add configurable minimum board amount
We want to have some minimum board amount that the server will
cosign for so they don't sign barely economical boards or anything
just above dust.

The default config is set to 20 000 sats which is similar to that
for minimum lightning channel capacity. Not much more to it than that.