bugfix(server): Client claims lightning receive
This fixes a flake I've observed in [CI](https://ci.2nd.dev/repos/6/pipeline/389/22)
that resulted in failed lightning payments for intra-ark payments.

In an Intra Ark payment a client will claim the outgoing HTLC by
revealing the pre-image. The server has to ensure it will make the
corresponding changes on the incoming HTLC to avoid losing any funds.

We use a tokio::broadcast channel for that communication. However,
it seems we aren't always holding a `Receiver` which might trigger
the channel to close.

This change ensures we keep always have at least one receiver listening