server: Fix race condition that allows double-spending vtxos
When using postgres we rely on a level of transaction isolation.
The default is READ_COMMITTED which we are using.

The core promise is that if you write in a transaction.
All readers will either see all of your changes or none.

The previous query relied on another property.
We first did a read to check if the Vtxo was spent.
If it wasn't, we would update the transaction.

This is buggy because other queries are allowed to update
the transaction in between.