bark/persist: Add preimage column to bark_lightning_send table
The `bark_pending_lightning_send` table is renamed to `bark_lightning_send`
to reflect the fact that we want to keep past finished lightning sends
in this table and not remove them anymore.

A preimage column is added to `bark_lightning_send` that will store
the preimage of a successful lightning payment.

A finished_at column is also introduced to indicate when the lightning
send moved out of a pending state. This is needed for failure cases
as the preimage may not be available, but the lightning send should
not show as pending anymore.

We also rename the persist model `PendingLightningSend` to `LightningSend`
to reflect that this model can represent a pending or non-pending
lightning send. An optional `preimage` field is also added to this and
is populated in the case of lightning sends that were successful.