lib: Make CheckpointedArkoorBuilder own the VTXO
I spend a few hours fighting lifetimes.
I promise you, it isn't worth it.

If you do this you have to worry about lifetimes a lot.
You'll see lifetime annotations everywhere.
I didn't manage to get it working, but the more I worked on it.
The more it clear it become. It isn't worth the effort.

You think we might save a few bytes by borrowing the `&Vtxo`?

On the server, this just isn't true.
We will get the full VTXO from the database, do some validation
and pass it to the builder who will hold it from their on.
We will not save a single clone. We'll just pollute all our
code with lifetimes.

For the client, it doesn't really matter either. It will
get VTXOs from the database, do some validation and pass them
to the database.

We now have a spurious clone when we create `CosignRequest` on the
clone. I'll promise, I will clean it up in one of the next commits