server: Implement database utilities
Implements some database utilities.

I have created a function which
- inserts new vtxos and can mark some (potentially other) vtxos as spend

I had to update the database trigger.
The `upsert_vtxos` and `mark_package_spent` are setting `updated_at` to
`NOW()`. I am executing them in the same `tx` and as a consequence
`NOW()` will be exactly the same. It is the timestamp at which the
transaction started. I made the `updated_at` check in the database
trigger a bit more lenient